app/ml_models/train_model.py
"""
Model training and evaluation utilities
"""

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score
import xgboost as xgb
import joblib
import logging

logger = logging.getLogger(__name__)

class ModelTrainer:
    """Handle model training"""
    
    def __init__(self, model_type='xgboost', **kwargs):
        self.model_type = model_type
        self.model = None
        self.params = kwargs
        
    def train(self, X_train, y_train, X_val=None, y_val=None):
        """Train the model"""
        # Implementation
        pass

class ModelEvaluator:
    """Evaluate model performance"""
    
    @staticmethod
    def calculate_metrics(y_true, y_pred) -> dict:
        """Calculate performance metrics"""
        # Implementation
        pass