requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0

# Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
tensorflow==2.13.0
scipy==1.10.1
pulp==2.7.0

# Database
sqlalchemy==2.0.21
sqlite3

# Frontend
streamlit==1.28.1
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2

# Utils
joblib==1.3.2
requests==2.31.0